Additional Experiments/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.44.0
datasets>=3.0.0
evaluate>=0.4.2
Pillow>=10.0.0
tqdm>=4.66.0
accelerate>=0.33.0

